requirement.txt
asgiref==3.7.2
bleach==6.0.0
cssmin==0.2.0
Django==3.2.20
django-appconf==1.0.5
django-compressor==4.4
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.18
django-summernote==0.8.20.0
jsmin==3.0.1
mysqlclient==2.2.0
Pillow==10.0.0
pytz==2023.3
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
webencodings==0.5.1
